chore(RiskAssessment): use long to prevent int stackoverflow when counting triggers, and order triggers alphabeticaly for better readability and maintenability

+++ RiskAssessmentService/src/main/java/org/medilabo/riskassessmentservice/model/RiskAssessment.java
@@ -7,6 +7,6 @@ public class RiskAssessment {
     Long patientId;
     String patientName;
     int age;
-    int triggerCount;
+    long triggerCount;
     RiskLevel riskLevel;
 }
